requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
azure-servicebus
azure-core-exceptions
azure-mgmt-servicebus
